refactored to make class more testable
